@@ -300,65 +297,67 @@
     "serve-favicon": "^2.5.0"
   },
   "devDependencies": {
-    "@babel/cli": "^7.0.0-beta.55",
-    "@babel/core": "^7.0.0-beta.55",
+    "@babel/cli": "^7.0.0",
+    "@babel/core": "^7.0.0",
     "babel-core": "^7.0.0-bridge.0",
-    "babel-eslint": "^8.2.6",
+    "babel-eslint": "^9.0.0",
     "babel-jest": "^23.4.2",
-    "babel-loader": "^8.0.0-beta.0",
-    "babel-plugin-istanbul": "^4.1.6",
-    "compression-webpack-plugin": "^1.1.11",
+    "babel-loader": "^8.0.0",
+    "babel-plugin-istanbul": "^5.0.1",
+    "compression-webpack-plugin": "^1.1.12",
     "coveralls": "^3.0.2",
     "css-loader": "^1.0.0",
-    "enzyme": "^3.3.0",
-    "enzyme-adapter-react-16": "^1.1.1",
-    "eslint": "^4.19.1",
-    "eslint-config-airbnb": "^17.0.0",
-    "eslint-config-prettier": "^2.9.0",
+    "enzyme": "^3.5.0",
+    "enzyme-adapter-react-16": "^1.3.1",
+    "eslint": "^5.3.0",
+    "eslint-config-airbnb": "^17.1.0",
+    "eslint-config-prettier": "^3.0.1",
     "eslint-import-resolver-webpack": "^0.10.1",
     "eslint-loader": "^2.1.0",
-    "eslint-plugin-import": "^2.12.0",
-    "eslint-plugin-jsx-a11y": "^6.0.3",
+    "eslint-plugin-import": "^2.14.0",
+    "eslint-plugin-jsx-a11y": "^6.1.1",
     "eslint-plugin-lodash": "^2.7.0",
     "eslint-plugin-prettier": "^2.6.2",
-    "eslint-plugin-react": "^7.9.1",
-    "file-loader": "^1.1.11",
+    "eslint-plugin-react": "^7.11.0",
+    "file-loader": "^2.0.0",
     "friendly-errors-webpack-plugin": "^1.7.0",
     "husky": "^0.14.3",
     "identity-obj-proxy": "^3.0.0",
     "imagemin-webpack-plugin": "^2.2.0",
-    "jest": "^23.4.2",
-    "lint-staged": "^7.2.0",
+    "jest": "^23.5.0",
+    "lint-staged": "^7.2.2",
     "lodash-webpack-plugin": "^0.11.5",
-    "mini-css-extract-plugin": "^0.4.1",
-    "nock": "^9.4.3",
-    "nodemon": "^1.18.3",
+    "mini-css-extract-plugin": "^0.4.2",
+    "nock": "^9.6.1",
+    "nodemon": "^1.18.4",
     "npm-run-all": "^4.1.3",
-    "optimize-css-assets-webpack-plugin": "^5.0.0",
-    "path": "^0.12.7",
-    "postcss": "^7.0.1",
-    "postcss-loader": "^2.1.6",
-    "prettier": "^1.14.0",
+    "optimize-css-assets-webpack-plugin": "^5.0.1",
+    "postcss": "^7.0.2",
+    "postcss-loader": "^3.0.0",
+    "prettier": "^1.14.2",
     "prettier-stylelint": "^0.4.2",
     "raf": "^3.4.0",
-    "react-test-renderer": "^16.4.1",
+    "react-test-renderer": "^16.4.2",
     "redux-mock-store": "^1.5.3",
-    "sass-loader": "^7.0.3",
+    "rimraf": "^2.6.2",
+    "sass-loader": "^7.1.0",
     "style-loader": "^0.22.0",
-    "stylelint": "^9.4.0",
+    "stylelint": "^9.5.0",
+    "stylelint-config-prettier": "^4.0.0",
     "stylelint-config-recommended-scss": "^3.2.0",
     "stylelint-config-standard": "^18.2.0",
-    "stylelint-scss": "^3.2.0",
+    "stylelint-scss": "^3.3.0",
     "stylelint-webpack-plugin": "^0.10.5",
-    "url-loader": "^1.0.1",
-    "webpack": "^4.16.3",
+    "uglifyjs-webpack-plugin": "^1.3.0",
+    "url-loader": "^1.1.1",
+    "webpack": "^4.17.1",
     "webpack-bundle-analyzer": "^2.13.1",
     "webpack-cli": "^3.1.0",
-    "webpack-dev-middleware": "^3.1.3",
-    "webpack-hot-middleware": "^2.22.3",
+    "webpack-dev-middleware": "^3.2.0",
+    "webpack-hot-middleware": "^2.23.0",
     "webpack-manifest-plugin": "^2.0.3"
   },
   "resolutions": {
-    "webpack-sources": "^1.1.0"
+    "webpack-sources": "^1.2.0"
   }
 }
