@@ -1,10 +1,8 @@
 /* Set your postcss-loader configuration here */
 
-const autoprefixer = require('autoprefixer');
-
 module.exports = {
   plugins: [
-    // add "-ms-" prefixes for Grid Layout
-    autoprefixer({ grid: true })
+    // Add "-ms-" prefixes for Grid Layout
+    require('autoprefixer')({ grid: true })
   ]
 };
