@@ -4,15 +4,19 @@ import { renderRoutes } from 'react-router-config';
 import Helmet from 'react-helmet';
 import { hot } from 'react-hot-loader';
 import { ToastContainer } from 'react-toastify';
+import { config as faConfig } from '@fortawesome/fontawesome-svg-core';
 
-import { Navigation } from 'components';
-import { EnsureLoggedIn, Login } from 'pages';
+import { Navigation } from '../components';
+import { EnsureLoggedIn, Login } from '../pages';
 
 import config from '../config';
 
 // Import your global styles here
+import '@fortawesome/fontawesome-svg-core/styles.css';
 import '../theme/bootstrap.scss';
 import './styles.scss';
+
+faConfig.autoAddCss = false;
 
 const scrollToTop = () => {
   if (typeof window === 'object') {
