@@ -2,7 +2,7 @@ import React, { PureComponent } from 'react';
 import { Redirect } from 'react-router-dom';
 import { connect } from 'react-redux';
 
-import { authActions } from 'actions';
+import { authActions } from '../../actions';
 
 class Logout extends PureComponent {
   componentDidMount() {
