@@ -2,7 +2,7 @@ import { routerMiddleware } from 'react-router-redux';
 import { createStore, applyMiddleware, compose } from 'redux';
 import thunk from 'redux-thunk';
 
-import rootReducer from 'reducers';
+import rootReducer from '../reducers';
 
 export default (history, initialState = {}) => {
   const middlewares = [
